fix: resolve path imports for vault router, cleanup client export from shared
