+++ shared/client.ts
@@ -1,1 +1,9 @@
-export * from '../worker/client'
+import { ClientRequestOptions, hc } from 'hono/client';
+import { DeadropWorkerApi } from '@api/client';
+
+export const createClient = (
+  url: string,
+  options?: ClientRequestOptions,
+) => hc<DeadropWorkerApi>(url, options);
+
+export type DeadropApiClient = ReturnType<typeof createClient>;
