+++ worker/client.ts
@@ -1,9 +1,1 @@
-import { hc, ClientRequestOptions } from 'hono/client';
-import { DeadropWorkerApi } from './src/app';
-
-export const createClient = (
-  url: string,
-  options?: ClientRequestOptions,
-) => hc<DeadropWorkerApi>(url, options);
-
-export type DeadropApiClient = ReturnType<typeof createClient>;
+export type { DeadropWorkerApi } from './src/app';
