+++ worker/src/routers/vault.ts
@@ -1,9 +1,9 @@
 import { zValidator } from '@hono/zod-validator';
 import { z } from 'zod';
 import { AppRouteParts } from '../constants';
-import { hono } from 'lib/http/core';
-import { createVaultUtils, vaultNameFromUserId } from 'lib/vault';
-import { authenticated, restricted } from 'lib/middleware';
+import { hono } from '../lib/http/core';
+import { createVaultUtils, vaultNameFromUserId } from '../lib/vault';
+import { authenticated, restricted } from '../lib/middleware';
 
 const VaultNameSchema = z.object({ name: z.string() });
 
@@ -79,7 +79,7 @@ const vaultRouter = hono()
   .get(
     AppRouteParts.NameParam,
     authenticated(),
-    zValidator('param', z.object({ name: z.string() })),
+    zValidator('param', VaultNameSchema),
     async (c) => {
       const userId = c.get('clerkAuth')!.userId!;
 
@@ -100,7 +100,7 @@ const vaultRouter = hono()
   .delete(
     AppRouteParts.NameParam,
     restricted(),
-    zValidator('param', z.object({ name: z.string() })),
+    zValidator('param', VaultNameSchema),
     async (c) => {
       const userId = c.get('clerkAuth')!.userId!;
